TicTacToeクラスのBoardに依存したメソッドをBoardのメソッドに変更
