+++ src/TicTacToe.java
@@ -1,5 +1,3 @@
-import java.util.Arrays;
-import java.util.stream.Stream;
 
 public class TicTacToe implements BoardGame {
   private TicTacToeBoard ticTacToeBoard;
@@ -30,41 +28,11 @@ public class TicTacToe implements BoardGame {
     boolean someoneWin;
     // TODO 盤面の状態から勝者を取得できるように修正
     int player = inputCount % 2 == 1 ? 1 : 2;
-    someoneWin = hasMatchLine(player) || hasMatchColumn(player) || hasMatchDiagonal(player);
+    someoneWin = ticTacToeBoard.hasMatchLine(player) || ticTacToeBoard.hasMatchColumn(player) || ticTacToeBoard.hasMatchDiagonal(player);
 
     return new Result(!someoneWin && inputCount < ticTacToeBoard.getLineSize() * ticTacToeBoard.getColumnSize()
         , someoneWin && inputCount % 2 == 1
         , someoneWin && inputCount % 2 == 0);
-  }
-
-  private boolean hasMatchDiagonal(int player) {
-    int[][] diagonalCells = new int[2][3];
-    // 斜線の値を取得する
-    for (int i = 0; i < ticTacToeBoard.getLineSize(); i++) {
-      diagonalCells[0][i] = ticTacToeBoard.getCellData(i, i);
-      diagonalCells[1][i] = ticTacToeBoard.getCellData(i, ticTacToeBoard.getColumnSize() - (i + 1));
-    }
-    return hasComplete(diagonalCells, player);
-  }
-
-  private boolean hasMatchColumn(int player) {
-    int[][] transposedBoard = new int[ticTacToeBoard.getColumnSize()][ticTacToeBoard.getLineSize()];
-    // 三目並べのボードに転置処理を行う
-    for (int i = 0; i < ticTacToeBoard.getLineSize(); i++) {
-      for (int j = 0; j < ticTacToeBoard.getColumnSize(); j++) {
-        transposedBoard[i][j] = ticTacToeBoard.getCellData(j, i);
-      }
-    }
-    return hasComplete(transposedBoard, player);
-  }
-
-  private boolean hasMatchLine(int player) {
-    return hasComplete(ticTacToeBoard.getBoard(), player);
-  }
-
-  private boolean hasComplete(int[][] board, int player) {
-    return Stream.of(board)
-        .anyMatch(line -> Arrays.stream(line).allMatch(cell -> cell == player));
   }
 
   /**
