+++ src/TicTacToeBoard.java
@@ -1,5 +1,6 @@
-import java.util.ArrayList;
-import java.util.List;
+
+import java.util.Arrays;
+import java.util.stream.Stream;
 
 public class TicTacToeBoard {
   private int lineSize;
@@ -30,5 +31,35 @@ public class TicTacToeBoard {
 
   public int[][] getBoard() {
     return board;
+  }
+
+  public boolean hasMatchDiagonal(int player) {
+    int[][] diagonalCells = new int[2][3];
+    // 斜線の値を取得する
+    for (int i = 0; i < lineSize; i++) {
+      diagonalCells[0][i] = board[i][i];
+      diagonalCells[1][i] = board[i][columnSize - (i + 1)];
+    }
+    return hasComplete(diagonalCells, player);
+  }
+
+  public boolean hasMatchColumn(int player) {
+    int[][] transposedBoard = new int[columnSize][lineSize];
+    // 三目並べのボードに転置処理を行う
+    for (int i = 0; i < lineSize; i++) {
+      for (int j = 0; j < columnSize; j++) {
+        transposedBoard[i][j] = board[j][i];
+      }
+    }
+    return hasComplete(transposedBoard, player);
+  }
+
+  public boolean hasMatchLine(int player) {
+    return hasComplete(board, player);
+  }
+
+  private boolean hasComplete(int[][] board, int player) {
+    return Stream.of(board)
+        .anyMatch(line -> Arrays.stream(line).allMatch(cell -> cell == player));
   }
 }
